+++ sdk/cosmos/azure-cosmos/azure/cosmos/_cosmos_client_connection.py
@@ -1226,7 +1226,7 @@ class CosmosClientConnection:  # pylint: disable=too-many-public-methods,too-man
         options: Optional[Mapping[str, Any]] = None,
         response_hook: Optional[Callable[[Mapping[str, Any], Mapping[str, Any]], None]] = None,
         **kwargs: Any
-    ) -> ItemPaged[dict[str, Any]]:
+    ) -> CosmosItemPaged:
         """Queries documents change feed in a collection.
 
         :param str collection_link: The link to the document collection.
@@ -1237,7 +1237,7 @@ class CosmosClientConnection:  # pylint: disable=too-many-public-methods,too-man
         :return:
             Query Iterable of Documents.
         :rtype:
-            query_iterable.QueryIterable
+            CosmosItemPaged
 
         """
 
@@ -1255,7 +1255,7 @@ class CosmosClientConnection:  # pylint: disable=too-many-public-methods,too-man
         partition_key_range_id: Optional[str] = None,
         response_hook: Optional[Callable[[Mapping[str, Any], Mapping[str, Any]], None]] = None,
         **kwargs: Any
-    ) -> ItemPaged[dict[str, Any]]:
+    ) -> CosmosItemPaged:
         """Queries change feed of a resource in a collection.
 
         :param str collection_link: The link to the document collection.
@@ -1267,7 +1267,7 @@ class CosmosClientConnection:  # pylint: disable=too-many-public-methods,too-man
         :return:
             Query Iterable of Documents.
         :rtype:
-            query_iterable.QueryIterable
+            CosmosItemPaged
 
         """
         if options is None:
@@ -1277,6 +1277,9 @@ class CosmosClientConnection:  # pylint: disable=too-many-public-methods,too-man
 
         path = base.GetPathFromLink(collection_link, http_constants.ResourceType.Document)
         collection_id = base.GetResourceIdOrFullNameFromLink(collection_link)
+
+        # Create shared list for thread-safe header capture
+        response_headers_list: list[CaseInsensitiveDict] = []
 
         def fetch_fn(options: Mapping[str, Any]) -> Tuple[list[dict[str, Any]], CaseInsensitiveDict]:
             if collection_link in self.__container_properties_cache:
@@ -1296,12 +1299,13 @@ class CosmosClientConnection:  # pylint: disable=too-many-public-methods,too-man
                 response_hook=response_hook,
                 **kwargs)
 
-        return ItemPaged(
+        return CosmosItemPaged(
             self,
             options,
             fetch_function=fetch_fn,
             collection_link=collection_link,
-            page_iterator_class=ChangeFeedIterable
+            page_iterator_class=ChangeFeedIterable,
+            response_headers_list=response_headers_list,
         )
 
     def _ReadPartitionKeyRanges(