+++ sdk/cosmos/azure-cosmos/azure/cosmos/aio/_container.py
@@ -985,7 +985,7 @@ class ContainerProxy:
             response_hook: Optional[Callable[[Mapping[str, str], dict[str, Any]], None]] = None,
             availability_strategy: Optional[Union[bool, dict[str, Any]]] = None,
             **kwargs: Any
-    ) -> AsyncItemPaged[dict[str, Any]]:
+    ) -> CosmosAsyncItemPaged:
         """Get a sorted list of items that were changed, in the order in which they were modified.
 
         :keyword int max_item_count: Max number of items to be returned in the enumeration operation.
@@ -1021,7 +1021,7 @@ class ContainerProxy:
         :keyword response_hook: A callable invoked with the response metadata.
         :paramtype response_hook: Callable[[Mapping[str, str], dict[str, Any]], None]
         :returns: An AsyncItemPaged of items (dicts).
-        :rtype: AsyncItemPaged[dict[str, Any]]
+        :rtype: ~azure.cosmos._cosmos_responses.CosmosAsyncItemPaged
         """
         ...
 
@@ -1037,7 +1037,7 @@ class ContainerProxy:
             response_hook: Optional[Callable[[Mapping[str, Any], dict[str, Any]], None]] = None,
             availability_strategy: Optional[Union[bool, dict[str, Any]]] = None,
             **kwargs: Any
-    ) -> AsyncItemPaged[dict[str, Any]]:
+    ) -> CosmosAsyncItemPaged:
         """Get a sorted list of items that were changed, in the order in which they were modified.
 
         :keyword dict[str, Any] feed_range: The feed range that is used to define the scope.
@@ -1069,7 +1069,7 @@ class ContainerProxy:
         :keyword response_hook: A callable invoked with the response metadata.
         :paramtype response_hook: Callable[[Mapping[str, str], dict[str, Any]], None]
         :returns: An AsyncItemPaged of items (dicts).
-        :rtype: AsyncItemPaged[dict[str, Any]]
+        :rtype: ~azure.cosmos._cosmos_responses.CosmosAsyncItemPaged
         """
         ...
 
@@ -1083,7 +1083,7 @@ class ContainerProxy:
             availability_strategy: Optional[Union[bool, dict[str, Any]]] = None,
             response_hook: Optional[Callable[[Mapping[str, Any], dict[str, Any]], None]] = None,
             **kwargs: Any
-    ) -> AsyncItemPaged[dict[str, Any]]:
+    ) -> CosmosAsyncItemPaged:
         """Get a sorted list of items that were changed, in the order in which they were modified.
 
         :keyword str continuation: The continuation token retrieved from previous response. It contains chang feed mode.
@@ -1105,7 +1105,7 @@ class ContainerProxy:
         :keyword response_hook: A callable invoked with the response metadata.
         :paramtype response_hook: Callable[[Mapping[str, str], dict[str, Any]], None]
         :returns: An AsyncItemPaged of items (dicts).
-        :rtype: AsyncItemPaged[dict[str, Any]]
+        :rtype: ~azure.cosmos._cosmos_responses.CosmosAsyncItemPaged
         """
         # pylint: enable=line-too-long
         ...
@@ -1121,7 +1121,7 @@ class ContainerProxy:
             availability_strategy: Optional[Union[bool, dict[str, Any]]] = None,
             response_hook: Optional[Callable[[Mapping[str, Any], dict[str, Any]], None]] = None,
             **kwargs: Any
-    ) -> AsyncItemPaged[dict[str, Any]]:
+    ) -> CosmosAsyncItemPaged:
         """Get a sorted list of items that were changed in the entire container,
          in the order in which they were modified.
 
@@ -1153,7 +1153,7 @@ class ContainerProxy:
         :keyword response_hook: A callable invoked with the response metadata.
         :paramtype response_hook: Callable[[Mapping[str, str], dict[str, Any]], None]
         :returns: An AsyncItemPaged of items (dicts).
-        :rtype: AsyncItemPaged[dict[str, Any]]
+        :rtype: ~azure.cosmos._cosmos_responses.CosmosAsyncItemPaged
         """
         ...
 
@@ -1161,7 +1161,7 @@ class ContainerProxy:
     def query_items_change_feed(  # pylint: disable=unused-argument
             self,
             **kwargs: Any
-    ) -> AsyncItemPaged[dict[str, Any]]:
+    ) -> CosmosAsyncItemPaged:
 
         """Get a sorted list of items that were changed, in the order in which they were modified.
 
@@ -1198,7 +1198,7 @@ class ContainerProxy:
         :keyword response_hook: A callable invoked with the response metadata.
         :paramtype response_hook: Callable[[Mapping[str, str], dict[str, Any]], None]
         :returns: An AsyncItemPaged of items (dicts).
-        :rtype: AsyncItemPaged[dict[str, Any]]
+        :rtype: ~azure.cosmos._cosmos_responses.CosmosAsyncItemPaged
         """
         # pylint: disable=too-many-statements
         validate_kwargs(kwargs)